pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "validators>=0.35.0",
]
package = ["page_analyzer"]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]
